Add files via upload
Sample script to map a json object of data to HubSpot form fields. The script takes 2 arguments: 1. An object of data and 2. An object of mappings. Given these two objects, the script will set these values into any fields on a form that match. This implementation fires change events on every field mapping due to HubSpot form rendering that uses …